# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Barberia",
        "description": "Servicio de API tipo SOA para implementación en una barberia",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/rol/nuevo": {
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Crear un nuevo registro de rol",
                "operationId": "cf70d952e82be1884a887d2d759b6cd5",
                "requestBody": {
                    "description": "Datos del rol para crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombreRol": {
                                        "description": "Nombre del rol",
                                        "type": "string",
                                        "example": "Admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se ha creado el registro de rol correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Los datos a validar no son correctos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Los datos a validar no son correctos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al crear el registro de rol",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al crear el registro de rol."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/rol/idRol/editar": {
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Actualizar un registro de rol específico",
                "operationId": "588f0f585e5a2dcfc68563784e757b4c",
                "parameters": [
                    {
                        "name": "idRol",
                        "in": "path",
                        "description": "ID del rol a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del rol para actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombreRol": {
                                        "description": "Nombre del rol",
                                        "type": "string",
                                        "example": "Admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El registro de rol ha sido actualizado correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Los datos a validar no son correctos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Los datos a validar no son correctos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al actualizar el registro de rol",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al actualizar el registro de rol."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/rol/idRol": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Mostrar un registro de rol específico",
                "operationId": "2b5dc8f765f28da1b9081147e3fb2a1a",
                "parameters": [
                    {
                        "name": "idRol",
                        "in": "path",
                        "description": "ID del rol a mostrar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "idRol": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "nombreRol": {
                                            "type": "string",
                                            "example": "Admin"
                                        },
                                        "created_at": {
                                            "description": "Fecha de creación del rol",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Fecha de última actualización del rol",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al mostrar el registro de rol",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al mostrar el registro de rol."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rol/idRol/eliminar": {
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Eliminar un registro de rol específico",
                "operationId": "d12a7a3baac8c2e158aac2692acf28fc",
                "parameters": [
                    {
                        "name": "idRol",
                        "in": "path",
                        "description": "ID del rol a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El registro de rol ha sido eliminado correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al eliminar el registro de rol",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al eliminar el registro de rol."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}